requirements.txt

Flask==2.3.2
firebase-admin==6.4.0
google-generativeai
tensorflow
numpy
